backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

langchain==0.1.4
langchain-community==0.0.17

pymilvus==2.3.3

neo4j==5.19.0

sentence-transformers==2.3.1

torch==2.1.2
torchvision==0.16.2
torchaudio==2.1.2

python-docx==1.1.0

requests==2.31.0

pydantic==2.5.3
python-dotenv==1.0.0
numpy==1.26.3

aiohttp==3.9.1
